Refactor code to modular programming
