+++ src/index.js
@@ -1,3 +1,6 @@
 import './styles/index.css';
 import './modules/render-list.js';
 import './modules/pop-up.js';
+import './modules/renderData.js';
+import './modules/addLike.js';
+import './modules/windowFunctions.js';
