+++ src/test/items_counter.test.js
@@ -1,0 +1,6 @@
+import itemsCounter from '../modules/itemCounter.js';
+// import  renderData  from '../modules/renderData.js';
+
+it('Store item in local storage', () => {
+  expect(1).not.toBeNull();
+});
